Tests/LibWeb/Text/expected/HTML/image-srcset-density-corrected-intrinsic-size.txt

currentSrc: 120.png?density=2
natural size: 60x60
layout size: 60x60
